requirements.txt
# Core runtime deps
runpod==1.7.0
torch>=2.2
torchaudio>=2.2
pyannote.audio>=3.1,<4
requests>=2.32
numpy>=1.26
soundfile>=0.12

# --- Optional dev / local testing (comment out in prod if you want smaller image) ---
# fastapi==0.115.0
# uvicorn[standard]==0.32.0
# pydantic>=2.8
